assignment/requirements.txt
apturl==0.5.2
asgiref==3.3.4
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
camelcase==0.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.7
django-import-export==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
itypes==1.2.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.0
monotonic==1.5
mysql-connector==2.2.9
mysql-connector-python==8.0.24
netifaces==0.10.4
num2words==0.5.10
numpy==1.20.2
oauth2client==4.1.2
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.7
pandas==1.0.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
SQLAlchemy==1.3.12
sqlparse==0.4.1
systemd-python==234
tablib==3.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
